requirements.txt
Flask>=1.0.0
cloudant==2.13.0
wiotp.sdk==0.11.0